Laying the groundwork and copy over the relevant bits from the old site
Putting down some initial structure. This uses a pedestal-esque Context Map
which gets passed around throughout the request. First it just contains the
request, then we add the necessary data, then the view adds itself (hiccup),
then that gets rendered and turned into a response. This way requests are
pipelines of composable bits.

The CSS has been copied over verbatim, although I was planning to use Garden on
this, but for now this is fine.

For the views I just took a random channel and date and took the currently
generated HTML, turned it into hiccup, and split it into a few functions. So far
all hard coded.

Data is being read from a couple of JSON and EDN files, this isn't necessarily
terrible but it does limit things a bit and means we might be keeping/loading
relatively large chunks in memory. I might just pump it all into postgress later
on.