Use reloaded.repl, also for th prod system.
If we're going to store the system map in production, we might as well store it
in the same place where we store it during dev-time, that way REPL helpers that
poke into the system map will work consistently.

Being able to start/stop in production might occasionally also prove handy.