Changes to how demo channels/users are exported
For the users we need to take into account that the :message/day field is three
hours off (for historical reasons, and not breaking existing links), so if we
only take the users of these specific days then those that only posted at the
very start of the first day will be missing, so add those that posted on 31
January.

Also add users who created channels, since we need to reference those.

Export :channel/creator as a lookup ref, rather than an entity map.

Remove :db/id, ids will be assigned on import.