Ignore direct messages
A direct message looks identical to a regular channel message, but the channel
id starts with a D rather than with a C. Since we don't have data about these
"direct message channels" Datomic will complain when using them in a lookup-ref.

The logbot is a silent bot, it does not talk back, and does not care what you
have to tell it. (Or maybe it's just a really good listener? We do not guarantee
your secrets are safe with us.)

Conclusion: drop direct messages on the floor.