Adds day navigation links to the header
Previous, the user must visit a page and pick from the list of days where
messages are available for that channel. The newly added navigation links allows
the user to navigate directly to the next or previous available day, without
leaving the message display page.

Note that the links will only be shown if it's possible to navigate to the next
or previous day.