Add test infrastructure, make username unique
This adds two edn files with test fixture data, as well as the REPL code to
generate them (re-run when the schema changes), and a `test-db` test helper for
loading them.

I added a few tests to demostrate their usage.

I tried to use :user/username in a lookup ref, and it turns out that it doesn't
actually have an index. I'm pretty sure this has to be unique, so marked it as
such, and in the process changed the user schema to look similar to the rest of
the schema data.