Add view test infrastructure and example
Add some helpers so it's easy to express view tests as a combination of an
enlive selector and expected hiccup.

Also include scope-capture during dev time, good to have that available, and
replace `add-dependency` with a macro, so you don't need to quote the dependency
vector.

While view tests are perhaps less important than tests at the query or route
level, there are certainly useful things to test here, in particular since we've
accrued a lot of logic in the views already.