Add logging Python script
This is the script that does the actual logging, i.e. listening for events and
spitting out log files. It's a simplified version of the rtmbot.py that's part
of the original slack-archivist, adapted to be a single script, instead of the
original "plugin" system.

(see https://bitbucket.org/ul/slack-archivist /
https://github.com/plexus/slack-archivist)

Included in the repo here for completeness. We'll keep this bit in Python
because it works, it boots quickly, and it has low requirements, making it easy
to run it in multiple places to provide some redundance. Right now whenever the
server reboots we lose several minutes of messages, going forward we'll try to
prevent that.

This adapted version also differs in an important way from the version we've
been using so far: instead of only logging "message" events, it logs everything
except a few noisy and not very useful events: "pong", "user_typing" and
"hello".

This will give us some new possibilities to consider in the future, e.g. keeping
a user's data up to date automatically based on "user_change" events, display
"reactions" based on "reaction_added" events, stay up to date with custom
emoji ("emoji_changed"), file uploads, new channels, etc.