Support channel renames
Channel renames were breaking the import, before importing new data we sync
channel data from slack, but both the slack-id and the channel name are consider
identities in datomic, and so trying to update a datom with the same slack-id
but different channel name caused datomic to look up one, and then complain
about the other one conflicting.

This change looks up the actual :db/id for those channels we already know about,
so that new ones get inserted, and renamed one get their new name applied.

This does mean that the log will only show the new name on old messages, not
necessarily the name that was present when the messages were posted.