Import threaded messages into the day of the thread
The concrete issue this solves is that we were seeing days in the listing for a
channel, that upon inspection didn't have any messages at all, because the
messages for those days were part of a thread that started on another day.

The :message/day field is really a denormalization to quickly determine on which
page to show a message, and so it makes sense to give threaded messages the
right :message/day immediately. This should allow simplifying some more queries,
this isn't yet part of this commit.