Better deal with recursive markup
We had some cases where an `:undecorated` tag was making its way through to the
hiccup compiler, and causing havoc. This adjusts the code that converts parsed
messages to hiccup to deal with arbitrary nesting.

It also improves message to text rendering (used for the the description meta
tag) in a similar way, and implements ~strikethrough~.

Deal more gracefully with missing user names. This is especially useful should
we end up parsing something as a username when it actually isn't.

Add a regression test case, and get rid of some :refer :all usage in the tests.