Reactmoji: fix bulk import, deal with reaction_removed
Implement reaction_removed, we were pretty close already. Make sure the bulk
import is able to handle reaction emoji events. There are several places where
we need to know message/key to be able to partition properly, and in this case
the lookup is via reaction/message.

Also clean up the comment section in the .repl namespace a bit.

Add a data-message-key to the HTML, so it's easy to reconcile what is shown on
the screen with what you can find in the database.